BackEnd/requirements.txt
Flask
mysql.connector
pymysql
gunicorn
flask_login
pytz
flask_cors